automaton_exp2/python/requirements.txt
antlr4-python3-runtime==4.12.0